Refactor route imports to use config files for better organization and clarity
