+++ unstable_rsc-parcel/src/routes/config.ts
@@ -2,24 +2,24 @@
 
 import type { unstable_RSCRouteConfig as RSCRouteConfig } from "react-router";
 
-import "./entry.client";
+import "../entry.client";
 
 export function routes() {
   return [
     {
       id: "root",
       path: "",
-      lazy: () => import("./routes/root/route"),
+      lazy: () => import("./root/route"),
       children: [
         {
           id: "home",
           index: true,
-          lazy: () => import("./routes/home/route"),
+          lazy: () => import("./home/route"),
         },
         {
           id: "about",
           path: "about",
-          lazy: () => import("./routes/about/route"),
+          lazy: () => import("./about/route"),
         },
       ],
     },
