+++ unstable_rsc-vite/src/entry.rsc.tsx
@@ -8,7 +8,7 @@ import {
 } from "@vitejs/plugin-rsc/rsc";
 import { unstable_matchRSCServerRequest as matchRSCServerRequest } from "react-router";
 
-import { routes } from "./routes";
+import { routes } from "./routes/config";
 
 function fetchServer(request: Request) {
   return matchRSCServerRequest({
