+++ unstable_rsc-vite/src/routes/config.ts
@@ -5,17 +5,17 @@ export function routes() {
     {
       id: "root",
       path: "",
-      lazy: () => import("./routes/root/route"),
+      lazy: () => import("./root/route"),
       children: [
         {
           id: "home",
           index: true,
-          lazy: () => import("./routes/home/route"),
+          lazy: () => import("./home/route"),
         },
         {
           id: "about",
           path: "about",
-          lazy: () => import("./routes/about/route"),
+          lazy: () => import("./about/route"),
         },
       ],
     },
